refactor : Feedback 조회할때 필요한 내용만 출력
