+++ backend/src/main/java/com/techeer/backend/api/feedback/dto/Response/FeedbackListResponse.java
@@ -1,4 +1,4 @@
-package com.techeer.backend.api.feedback.dto;
+package com.techeer.backend.api.feedback.dto.Response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -13,14 +13,14 @@ public class FeedbackListResponse {
 
     private Long resumeId;
 
-    private List<Feedback> feedbacks;
+    private List<FeedbackResponse> feedbacks;
 
-    private FeedbackListResponse(Long resumeId, List<Feedback> feedbacks){
+    private FeedbackListResponse(Long resumeId, List<FeedbackResponse> feedbacks){
         this.resumeId = resumeId;
         this.feedbacks = feedbacks;
     }
 
-    public static  FeedbackListResponse from(Long resumeId, List<Feedback> feedbacks){
+    public static  FeedbackListResponse from(Long resumeId, List<FeedbackResponse> feedbacks){
         return new FeedbackListResponse(resumeId, feedbacks);
     }
 }
