+++ backend/src/main/java/com/techeer/backend/api/resume/dto/response/FetchResumeContentResponse.java
@@ -4,6 +4,7 @@ package com.techeer.backend.api.resume.dto.response;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import com.techeer.backend.api.feedback.domain.Feedback;
+import com.techeer.backend.api.feedback.dto.Response.FeedbackResponse;
 import com.techeer.backend.api.resume.domain.Position;
 import com.techeer.backend.api.resume.domain.Resume;
 import com.techeer.backend.api.resume.domain.ResumeTechStack;
@@ -21,10 +22,10 @@ public class FetchResumeContentResponse {
     private int career;
     private List<ResumeTechStack> techStack;
     private String fileUrl;
-    private List<Feedback> feedbacks;
+    private List<FeedbackResponse> feedbacks;
 
     private FetchResumeContentResponse(Long id, String username, Position position, int career,
-                                       List<ResumeTechStack> techStack, String fileUrl, List<Feedback> feedbacks) {
+                                       List<ResumeTechStack> techStack, String fileUrl, List<FeedbackResponse> feedbacks) {
         this.resumeId = id;
         this.userName = username;
         this.position = position;
@@ -34,7 +35,7 @@ public class FetchResumeContentResponse {
         this.feedbacks = feedbacks;
     }
 
-    public static FetchResumeContentResponse from(Resume resume, List<Feedback> feedbacks){
+    public static FetchResumeContentResponse from(Resume resume, List<FeedbackResponse> feedbacks){
         return new FetchResumeContentResponse(resume.getId(),
                 resume.getUsername(), resume.getPosition(), resume.getCareer(), resume.getResumeTechStacks(),
                 resume.getUrl(), feedbacks);
